ResultTranslator as inner class in IResult
